Fix type definitions for posts and campaigns
Previously result of infiniteQueries was not properly typed.
